refactor: split four oversized SDL modules into re-export packages
Convert raes/runtime_security_monitoring, raes/semantics/objective_semantics,
raes/semantics/participant_behavior, and raes_operations/_evidence_run_artifact
from single modules over the ADR-015 500-line cap into packages behind
API-stable __init__ re-export shims (issue #51). No behavior change; every
external import line is unchanged.

- objective_semantics: partition_objective_dependencies and
  analyze_objective_semantics stay on the package facade so the
  OBJECTIVE_*_DEPENDENCY_ROLES constants the test monkeypatches on the module
  remain the ones partitioning reads.
- runtime_security_monitoring: preserve the exact ordered __all__ and the
  re-exported detection-definition names; make the runtime family-invariant
  discovery (shared-helper glob and enum discovery) walk package submodules so
  the invariants keep enforcing across the split.
- Drain the four oversized-allowlist entries and repoint the SDL catalog and
  semantic-coverage docs at the new facade __init__ paths.
