+++ implementations/python/packages/raes/semantics/participant_behavior/__init__.py
@@ -1,0 +1,10 @@
+"""Name-level participant behavior semantics (SEM-208/209/210)."""
+
+from __future__ import annotations
+
+from ._analysis import analyze_participant_behavior
+from ._types import (
+    ParticipantBehaviorAnalysis,
+    ParticipantBehaviorIssue,
+    ParticipantBehaviorReference,
+)
