+++ implementations/python/pyproject.toml
@@ -182,4 +182,6 @@ ignore = [
 "packages/raes_runtime/control_plane_api/__init__.py" = ["F401"]  # intentional facade re-exports (_receipt_response)
 "packages/raes_mcp/tools/authoring/__init__.py" = ["F401"]  # intentional facade re-export (register)
 "packages/raes_mcp/tools/inspection/__init__.py" = ["F401"]  # intentional facade re-export (register)
+"packages/raes/semantics/objective_semantics/__init__.py" = ["F401"]  # intentional facade re-exports
+"packages/raes/semantics/participant_behavior/__init__.py" = ["F401"]  # intentional facade re-exports
 "packages/raes_contracts/realization_envelope.py" = ["E402"]  # late import breaks a carrier dependency cycle
